CORE-1717: update the active plan endpoint to return all usage and quota information
